Remove old model Organization, which is now in the db package.
Created organization query/mutation resolvers.
